refactor some vibes
