+++ components/admin-notes/AdminNoteView.tsx
@@ -1,7 +1,11 @@
 'use client';
 
-import { AdminNoteDTO, Roles } from '@osu-tournament-rating/otr-api-client';
-import { StickyNote } from 'lucide-react';
+import {
+  AdminNoteDTO,
+  AdminNoteRouteTarget,
+  Roles,
+} from '@osu-tournament-rating/otr-api-client';
+import { Loader2, StickyNote } from 'lucide-react';
 import { useSession } from 'next-auth/react';
 import { useState } from 'react';
 import { Button } from '../ui/button';
@@ -13,40 +17,93 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '../ui/dialog';
-import AdminNoteForm, { AdminNoteFormProps } from './AdminNoteForm';
-import AdminNotesList from './AdminNoteList';
 import { AdminNoteRouteTargetEnumHelper } from '@/lib/enums';
+import { adminNoteFormSchema } from '@/lib/schema';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { useForm } from 'react-hook-form';
+import { createNote } from '@/lib/actions/admin-notes';
+import { toast } from 'sonner';
+import {
+  Form,
+  FormField,
+  FormItem,
+  FormControl,
+  FormMessage,
+} from '../ui/form';
+import { Textarea } from '../ui/textarea';
+import { z } from 'zod';
+import AdminNotesList from './AdminNoteList';
+
+interface AdminNoteViewProps {
+  /**
+   * Admin notes
+   */
+  notes: AdminNoteDTO[];
+
+  /**
+   * Type of parent entity
+   */
+  entity: AdminNoteRouteTarget;
+
+  /**
+   * Id of the parent entity
+   */
+  entityId: number;
+
+  /**
+   * Optional display name for the parent entity. Uses entityId if not provided
+   */
+  entityDisplayName?: string;
+}
+
 export default function AdminNoteView({
   notes,
-  props,
-}: {
-  notes: AdminNoteDTO[];
-  props: AdminNoteFormProps;
-}) {
-  const [showNotification, setShowNotification] = useState(true);
+  entity,
+  entityId,
+  entityDisplayName,
+}: AdminNoteViewProps) {
   const { data: session } = useSession();
 
+  const form = useForm<z.infer<typeof adminNoteFormSchema>>({
+    resolver: zodResolver(adminNoteFormSchema),
+    defaultValues: {
+      note: '',
+    },
+  });
+
+  const [showNotification, setShowNotification] = useState(true);
   const notify = !!notes.length;
 
   if (!session?.user?.scopes?.includes(Roles.Admin) && !notify) {
     return null;
   }
 
-  const entityMetadata = AdminNoteRouteTargetEnumHelper.getMetadata(
-    props.entity
-  );
+  const entityMetadata = AdminNoteRouteTargetEnumHelper.getMetadata(entity);
+  entityDisplayName ??= `${entityMetadata.text} ${entityId}`;
 
-  const handleClick = () => {
-    if (showNotification) {
-      setShowNotification(false);
+  async function onSubmit(data: z.infer<typeof adminNoteFormSchema>) {
+    try {
+      await createNote({
+        entityId: entityId,
+        entity: entity,
+        body: data.note,
+      });
+
+      toast.success(`Created admin note for ${entityDisplayName}`);
+    } catch {
+      toast.error(`Failed to create admin note for ${entityDisplayName}`);
     }
-  };
+  }
 
   return (
     <Dialog>
       <DialogTrigger asChild>
         <Button
-          onClick={handleClick}
+          onClick={() => {
+            if (showNotification) {
+              setShowNotification(false);
+            }
+          }}
           className="relative h-5 w-5"
           variant={'ghost'}
         >
@@ -66,13 +123,56 @@ export default function AdminNoteView({
           <DialogTitle>Admin Notes</DialogTitle>
           <DialogDescription>
             Viewing admin notes for{' '}
-            <span className="font-semibold">
-              {entityMetadata.text} {props.entityId}
-            </span>
+            <span className="font-semibold">{entityDisplayName}</span>
           </DialogDescription>
         </DialogHeader>
-        <AdminNoteForm {...props} />
-        <AdminNotesList {...props} notes={notes} />
+
+        {/* New note creation form */}
+        <Form {...form}>
+          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-2">
+            <FormField
+              control={form.control}
+              name="note"
+              render={({ field }) => (
+                <FormItem>
+                  <FormControl>
+                    <Textarea
+                      placeholder="Add your note here"
+                      className="min-h-24 resize-none"
+                      {...field}
+                    />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+            <div className="flex justify-between space-x-2">
+              {/* Reset changes */}
+              <Button
+                type="reset"
+                variant={'secondary'}
+                onClick={() => form.reset()}
+                disabled={
+                  !form.formState.isDirty || form.formState.isSubmitting
+                }
+              >
+                Discard
+              </Button>
+              {/* Save changes */}
+              <Button
+                type="submit"
+                disabled={!form.formState.isValid || !form.formState.isDirty}
+              >
+                {form.formState.isSubmitting ? (
+                  <Loader2 className="animate-spin" />
+                ) : (
+                  'Submit'
+                )}
+              </Button>
+            </div>
+          </form>
+        </Form>
+        <AdminNotesList entity={entity} notes={notes} />
       </DialogContent>
     </Dialog>
   );
